requirements.txt
bertopic==0.16.0
hdbscan==0.8.33
langchain==0.1.14
langchain_community==0.0.31
langchain_core==0.1.38
langchain_text_splitters==0.0.1
nltk==3.8.1
PyPDF2==2.10.5
PyPDF2==3.0.1
streamlit==1.32.2
umap_learn==0.5.5
